+++ src/elf_helpers.h
