+++ src/user_helpers.h
